all int's are changed to unsigned long's to avoid overflow for 600851475143. Tested for 13195, sucessful.

+++ p3/c/p3.c
@@ -3,16 +3,16 @@
 #include <math.h>
 
 /* INITIALIZE NUMBER TO BE FACTORIZED */
-int num = 13195;
+unsigned long num = 13195;
 
 /* FUNCTIONS DECLARATIONS */
-int eliminate_prime(int, int);
+unsigned long eliminate_prime(unsigned long, unsigned long);
 
 /* MAIN PROGRAM */
 int main(void)
 {
-	int largest = 1;
-	int my_num = num;
+	unsigned long largest = 1;
+	unsigned long my_num = num;
 
 	/* Remove 2s from the number so that it becomes odd */
 	if (num%2 == 0)
@@ -28,8 +28,8 @@ int main(void)
 	 * Thus, my_num == 3 and my_num == 5 cases must be handled m
 	 * anually
 	 */
-	int i;
-	int sqrt_num = sqrt(my_num);
+	unsigned long i;
+	unsigned long sqrt_num = sqrt(my_num);
 	for(i=3; i <= sqrt_num; i++)
 	{
 		if (my_num % i == 0)
@@ -44,13 +44,13 @@ int main(void)
 	if (largest == 1)
 		largest = num;
 
-	printf("The largest prime number that divides %d is %d\n", num, largest);
+	printf("The largest prime number that divides %lu is %lu\n", num, largest);
 
 	return 0;
 }
 
 
-int eliminate_prime(int x, int y)
+unsigned long eliminate_prime(unsigned long x, unsigned long y)
 {
 	/* This function divides x to y as long as x is divisible by y.
 	 * Therefore, the returned value is not divisible by y anymore.
